+++ APIBlox.AspNetCore/Types/FilteredOrderedQuery.cs
@@ -1,4 +1,5 @@
-﻿using APIBlox.AspNetCore.Contracts;
+﻿using APIBlox.AspNetCore.Attributes;
+using APIBlox.AspNetCore.Contracts;
 using APIBlox.NetCore.Extensions;
 using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Mvc;
@@ -9,26 +10,37 @@ namespace APIBlox.AspNetCore.Types
     ///     Class FilteredOrderedQuery.
     ///     Implements the <see cref="APIBlox.AspNetCore.Types.FilteredQuery" />
     ///     Implements the <see cref="APIBlox.AspNetCore.Contracts.IOrderedQuery" />
+    ///     <para>
+    ///         Be sure to also call the AddFromQueryWithAlternateNamesBinder Mvc/MvcCore
+    ///         builder extension method to allow alternate names to be used.
+    ///     </para>
+    ///     <para>
+    ///         Alternates In addition to FilteredQuery, OrderBy = $OrderBy, SortBy, $SortBy, Sort, $Sort
+    ///     </para>
     /// </summary>
     /// <seealso cref="APIBlox.AspNetCore.Types.FilteredQuery" />
     /// <seealso cref="APIBlox.AspNetCore.Contracts.IOrderedQuery" />
     public class FilteredOrderedQuery : FilteredQuery, IOrderedQuery
     {
+        #region -    Constructors    -
+
         /// <summary>
         ///     Initializes a new instance of the <see cref="FilteredProjectedPaginationQuery" /> class.
         /// </summary>
         /// <inheritdoc />
         public FilteredOrderedQuery()
         {
-            Map.TryAdd("OrderBy", new[] {"$OrderBy", "SortBy", "$SortBy", "Sort", "$Sort"});
+            Map.TryAdd("OrderBy", new[] { "$OrderBy", "SortBy", "$SortBy", "Sort", "$Sort" });
         }
+
+        #endregion
 
         /// <summary>
         ///     Sets the order by.  Usage is determined by the API itself, please seek external documentation.
         /// </summary>
         /// <value>The order by.</value>
         /// <inheritdoc />
-        [FromQuery(Name = "orderBy")]
+        [FromQueryWithAlternateNames(new[] { "orderBy", "$OrderBy", "SortBy", "$SortBy", "Sort", "$Sort" })]
         public string OrderBy { get; set; }
 
         /// <summary>