+++ APIBlox.AspNetCore/Types/OrderedQuery.cs
@@ -1,4 +1,5 @@
-﻿using APIBlox.AspNetCore.Contracts;
+﻿using APIBlox.AspNetCore.Attributes;
+using APIBlox.AspNetCore.Contracts;
 using APIBlox.NetCore.Extensions;
 using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Mvc;
@@ -9,6 +10,13 @@ namespace APIBlox.AspNetCore.Types
     ///     Class OrderedQuery.
     ///     Implements the <see cref="APIBlox.AspNetCore.Types.Query" />
     ///     Implements the <see cref="APIBlox.AspNetCore.Contracts.IOrderedQuery" />
+    ///     <para>
+    ///         Be sure to also call the AddFromQueryWithAlternateNamesBinder Mvc/MvcCore
+    ///         builder extension method to allow alternate names to be used.
+    ///     </para>
+    ///     <para>
+    ///         Alternates to OrderBy = $OrderBy, SortBy, $SortBy, Sort, $Sort
+    ///     </para>
     /// </summary>
     /// <seealso cref="APIBlox.AspNetCore.Types.Query" />
     /// <seealso cref="APIBlox.AspNetCore.Contracts.IOrderedQuery" />
@@ -19,7 +27,7 @@ namespace APIBlox.AspNetCore.Types
         /// </summary>
         public OrderedQuery()
         {
-            Map.TryAdd("OrderBy", new[] {"$OrderBy", "SortBy", "$SortBy", "Sort", "$Sort"});
+            Map.TryAdd("OrderBy", new[] { "$OrderBy", "SortBy", "$SortBy", "Sort", "$Sort" });
         }
 
         /// <inheritdoc />
@@ -27,7 +35,7 @@ namespace APIBlox.AspNetCore.Types
         ///     Sets the order by.  Usage is determined by the API itself, please seek external documentation.
         /// </summary>
         /// <value>The order by.</value>
-        [FromQuery(Name = "orderBy")]
+        [FromQueryWithAlternateNames(new[] { "orderBy", "$OrderBy", "SortBy", "$SortBy", "Sort", "$Sort" })]
         public string OrderBy { get; set; }
 
         /// <summary>