+++ APIBlox.AspNetCore/Types/ProjectedQuery.cs
@@ -1,15 +1,23 @@
-﻿using APIBlox.AspNetCore.Contracts;
+﻿using APIBlox.AspNetCore.Attributes;
+using APIBlox.AspNetCore.Contracts;
 using APIBlox.NetCore.Extensions;
 using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Mvc;
 
 namespace APIBlox.AspNetCore.Types
 {
-    /// <inheritdoc cref="Query" />
     /// <summary>
     ///     Class ProjectedQuery.
+    /// Implements the <see cref="APIBlox.AspNetCore.Types.Query" />
+    /// Implements the <see cref="APIBlox.AspNetCore.Contracts.IProjectedQuery" />
+    ///     <para>
+    ///         Be sure to also call the AddFromQueryWithAlternateNamesBinder Mvc/MvcCore
+    ///         builder extension method to allow alternate names to be used.
+    ///     </para>
+    ///     <para>
+    ///         Alternates to Select = $Select, Project, $Project
+    ///     </para>
     /// </summary>
-    /// <seealso cref="T:APIBlox.AspNetCore.Types.OrderedQuery" />
     public class ProjectedQuery : Query, IProjectedQuery
     {
         /// <inheritdoc />
@@ -18,7 +26,7 @@ namespace APIBlox.AspNetCore.Types
         /// </summary>
         public ProjectedQuery()
         {
-            Map.TryAdd("Select", new[] {"$Select", "Project", "$Project"});
+            Map.TryAdd("Select", new[] { "$Select", "Project", "$Project" });
         }
 
         /// <inheritdoc />
@@ -27,6 +35,7 @@ namespace APIBlox.AspNetCore.Types
         /// </summary>
         /// <value>The select.</value>
         [FromQuery(Name = "select")]
+        [FromQueryWithAlternateNames(new[] { "select", "$Select", "Project", "$Project" })]
         public string Select { get; set; }
 
         /// <inheritdoc />