+++ Examples/APIBlox Features/Examples.Features/Controllers/ExamplesController.cs
@@ -1,4 +1,6 @@
-﻿#if UseAPIBlox
+﻿
+using APIBlox.AspNetCore.Attributes;
+#if UseAPIBlox
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -47,7 +49,7 @@ namespace Examples.Controllers
         /// <param name="wait">if not null then simulate wait middleware will kick in.</param>
         /// <returns>ActionResult&lt;IEnumerable&lt;System.String&gt;&gt;.</returns>
         [HttpGet]
-        public ActionResult<IEnumerable<string>> Get([FromQuery] PaginationQuery query = null, string wait = null)
+        public ActionResult<IEnumerable<string>> Get([FromQuery] PaginationQuery query, string wait = null)
         {
             var examples = new List<string>();
 
@@ -64,7 +66,7 @@ namespace Examples.Controllers
         /// <param name="description">The description, when empty no error details are displayed</param>
         /// <returns>ActionResult.</returns>
         [HttpGet("problemResult")]
-    #if UseAPIBlox
+#if UseAPIBlox
         public ActionResult GetProblemResultExample(CommonStatusCodes statusCode = CommonStatusCodes.Status403Forbidden, string description = null)
         {
             if (statusCode == CommonStatusCodes.Status200Ok || statusCode == CommonStatusCodes.Status204NoContent)
@@ -74,13 +76,13 @@ namespace Examples.Controllers
             errObject.SetError(statusCode, description);
 
             return new ProblemResult(errObject);
-        #else
+#else
         public ActionResult GetErrorResponseExample(int statusCode, string description = null)
         {
             HttpContext.Response.StatusCode = statusCode;
 
             return new ObjectResult(description);
-        #endif
+#endif
         }
 
         /// <summary>
@@ -118,14 +120,14 @@ namespace Examples.Controllers
         {
             throw new Exception("Be sure to try this out in RELEASE mode"
 
-                //,
-                //new IndexOutOfRangeException("As most if not all of this",
-                //    new ArgumentException("error information is NOT displayed in production",
-                //        new FileNotFoundException("By the way here is your message",
-                //            new NullReferenceException(exceptionMessage)
-                //        )
-                //    )
-                //)
+            //,
+            //new IndexOutOfRangeException("As most if not all of this",
+            //    new ArgumentException("error information is NOT displayed in production",
+            //        new FileNotFoundException("By the way here is your message",
+            //            new NullReferenceException(exceptionMessage)
+            //        )
+            //    )
+            //)
             );
         }
 
@@ -134,11 +136,11 @@ namespace Examples.Controllers
         /// </summary>
         /// <param name="requestResource">The request resource.</param>
         [HttpPost("{valueId:int}/subResources")]
-    #if UseAPIBlox
+#if UseAPIBlox
         public ActionResult Post(ExampleRequestObject requestResource)
-    #else
+#else
         public ActionResult Post(ExampleRequestObject requestResource)
-    #endif
+#endif
         {
             //
             //  SIDE NOTE:
@@ -146,9 +148,9 @@ namespace Examples.Controllers
             // I'm lazy and that's not the point of all this... :/
 
             return Conflict(new
-                {
-                    detail = "Please see errors property for more details",
-                    errors = new[]
+            {
+                detail = "Please see errors property for more details",
+                errors = new[]
                     {
                         new
                         {
@@ -156,7 +158,7 @@ namespace Examples.Controllers
                             title = "The request method does not allow this functionality as upsert semantics are not supported."
                         }
                     }
-                }
+            }
             );
 
             //  return Ok(new {Id = 1, requestResource.CoolNewValue, requestResource.ValueId});